+++ api/routes/api.js
@@ -147,7 +147,8 @@ const gerUsernameRegexpForSearch = (gamertag) => {
 }
 
 router.get('/config', function(req, res, next) {
-  console.log("####### API SESSION", req.session);
+  // console.log("####### API SESSION", req.session);
+  // console.log("####### API COOKIES", req.cookies);
   const lol_regions_short = logic_lol.get_regions_short();
   res.status(200).json({
     platforms: config.supported_platforms,